ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
redis==5.0.1
asyncpg==0.29.0
loguru==0.7.2
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
joblib==1.3.2
prometheus-client==0.19.0
pydantic==2.5.2
pydantic-settings==2.1.0
aiohttp==3.9.1
APScheduler==3.10.4
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Testing dependencies
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2